[models/instance] serialize video_size_limit as a long my friendica returned 4194304000 which is way above the 32-bit int limit

+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/model/InstanceV1.kt
@@ -89,7 +89,7 @@ data class InstanceV1(
 			@SerialName("image_matrix_limit")
 			val imageMatrixLimit: Int = 0,
 			@SerialName("video_size_limit")
-			val videoSizeLimit: Int = 0,
+			val videoSizeLimit: Long = 0,
 			@SerialName("video_frame_limit")
 			val videoFrameLimit: Int = 0,
 			@SerialName("video_matrix_limit")
